leaftlet map with markers and custom popup
Summary:
* adds leaflet map support
* adds markers and custom PL popups

There was an issue where the background image of `.post-popup.has-image` was being seen on the map, so I added `overflow: hidden;` to `.post-popup` which hid the `:after` arrow, but Leaflet provides `:after` arrows, so I made that call. I wrote a bit of JS to style the white arrow black on `.has-image` but although it works in the browser console, it is not working in the js file (I have no idea why). Also Leaflet programatically calculates the position of the pop up, so the PL `:after` arrow was offset. I thought going with defaults was the easiest choice, since it still conveys our design.
Fixes T1358

Test Plan:
* load "Blocks"
* click markers
* test map
* review code

Reviewers: sophshep

Reviewed By: sophshep

Subscribers: brandonrosage

Maniphest Tasks: T1358

Differential Revision: https://phabricator.ushahidi.com/D703