undo tab dropdown
Summary:
Got it working on home-list page

Getting it to work on all the other pages

This ended up being a lot more complicated than I expected it to be. Especially on pages such as activity where it is sometimes shown, sometimes not. I plan on spending an hour or so tomorrow seeing if i can clean up the code at all (once I've had a break from it!)

Test Plan: Look at pattern on all layouts at different screen sizes, check it works okay.

Reviewers: seth

Reviewed By: seth

Subscribers: brandonrosage

Differential Revision: https://phabricator.ushahidi.com/D770